route controllers GET/products, GET/products/:id & DELETE/products/:id
