+++ api/src/routes/controllers/products.js
@@ -3,7 +3,7 @@ const {Product, ProductsCategory} = require('../../db');
 
 
 
-//PRODUCTS DB:
+//ALL PRODUCTS FROM DB:
 const allProducts = async() => {
     try {
         const dbProduct = await Product.findAll({
@@ -31,8 +31,7 @@ const allProducts = async() => {
     };
 }
 
-
-//RECIPES POR ID:
+//PRODUCTS BY ID:
 const productsId = async(idP) => {
     try {
         const totalProducts = await allProducts();
@@ -47,6 +46,63 @@ const productsId = async(idP) => {
     }
 }
 
+//--------------------------------------------------------------------------------------------------------------
+
+const getProducts = async(req, res) => {
+ 
+    try {
+        const qname = req.query.name;
+        const totalProducts = await allProducts();
+        ;
+        if(qname) {
+            const productsWithName = totalProducts.filter((r) => r.name.toLowerCase().includes(qname.toLowerCase()));
+            (productsWithName.length) ?
+            res.send(productsWithName)
+            : res.status(404).send('Producto no encontrada');            
+        } else {
+            res.send(totalProducts);
+        }
+
+    } catch (err) {
+        console.log(err);
+        res.status(404).send('Problemas en el controlador de la ruta GET/products');
+    }; 
+};
 
 
-module.exports = {allProducts, productsId};
+
+const getProductsId = async(req, res) => {
+    try {
+        const id = req.params.id;
+        
+        const result = await productsId(id);
+        if (result) {
+            return res.send(result);
+        } else {
+            res.status(404).send('Id no existente')
+        };
+    }
+    catch(err) {
+        res.status(404).send('Problemas en el controlador de la ruta GET/products/:id');
+    }
+   
+};
+
+
+
+const deleteProduct = async (req, res, next)=>{
+    const { id } = req.params;
+    try {
+      await  Product.destroy({
+        where:{
+            id:id,
+        }
+      })
+      res.status(200).send('Producto eliminado')
+    } catch (error) {
+        res.status(400).send('error')
+    }
+    
+}
+
+module.exports = {getProducts, getProductsId, deleteProduct};
